requirements.txt
# --- Data Ingestion & Manipulation ---
pybaseball==2.2.7
pandas==2.2.2
pyarrow==16.1.0  # for Parquet files

# --- Database ---
duckdb==0.10.2
duckdb-engine==0.11.0 # For easier integration with tools like Pandas

# --- API Layer (for later steps) ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0

# --- LLM Integration (for later steps) ---
openai==1.30.1    
deepseek==1.0.0
python-dotenv==1.0.1 

boto3==1.34.131
